backend/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.1

# Async Processing
celery==5.3.4

# Data Processing
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
scikit-learn==1.3.2
statsmodels==0.14.1

# Time Series
# Note: TimescaleDB is a PostgreSQL extension, installed in the database, not a Python package
ta==0.10.2

# Machine Learning (commented out for faster builds - uncomment when needed)
# tensorflow==2.15.0
# torch==2.1.1
# transformers==4.36.0

# Financial Data
yfinance==0.2.28
alpha-vantage==2.3.1
ccxt==4.1.77

# WebSocket
websockets==12.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Monitoring
prometheus-client==0.19.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.12.0
httpx==0.25.0
factory-boy==3.3.0

# Development
black==23.11.0
flake8==6.1.0
mypy==1.7.1
isort==5.12.0

# Production utilities
aiohttp==3.9.1
slowapi==0.1.9
